Dev/robin/9236 merklesealer paging unreliable (#102)
* defer reading blobs and verifying seals until we know the log has progressed

The fundamental cause of AB#9236 was egregious ineficiency. It was
taking many seconds, some times 10's of seconds to process the tenants
that *might* need confirming.

This change re-arrangest things so we defer the massif reading entirely
until we know the log has progressed.

We do this by relying on the content length and the last modified
properties we get from azure blob storage. At worst, failure simply
causes us not to seal the log. It can not result in corruption or mis
signing.

Before actually creating a new seal we, as before, verify the current
seal and verify the new log state is consistent with the verified previous
state.

The methods for navigating the mmrblob basic format were previously only
available on the MassifContext. To facilitate checking the number of
entries in the log based only on the content length those have been
re-factored as stand alone functions which now live in the logformat.go
file

Also renamed from logentry.go to logformat.go

In summary,

* Use the last modified on the blob meta data to decide if the
  confirmation interval has elapsed rather than reading the whole massif
  of every tenant regardless.
* Use the content length of the masssif to decide if more entries have
  been added since the last seal was made
* Make the fundamental mmrblob format functions available as module
  scoped functions.
* Defer verification of the seal signature until *just* after the length
  check. This fails safe and is measurably a lot more efficient.
* Re-organise the pending tenant checks so we don't read the massif's
  and seals multiple times from blob storage

restore cheatsheet bad merge

fixup linter issues

fixup tests

* use lazy context rather than the more obscure lazy context

* review updates

* logical context -> lazy context
* remove the TODO: add tests that are already filled in

---------

Co-authored-by: Robin Bryce <robin.bryce@datatrails.ai>