fix: make tags accessible to callers of signerootreader
A quirk of the interface for lazily reading the root seal blobs
prevented the blob tags from being available to the caller of
ReadLogicalContext.

This also now works consistently for all methods on the signed root
reader which use this method internaly.

Rather than egregiously break an existing public interface, and also
because it seems cleaner, we just save the last context read and provide
a method to retrieve it if wanted.

The method to retrieve will error if the context hasn't been read at
all.

AB#9570