fix: replciation issue local log and seal corruption
When replicating the remote log file and the remote seal to local disc
the files must be opened in O_TRUNC mode.

The single corner case where this happens required file system based
tests and we only have those in veracity.

Follow on work required if we want those tests to exist in this repo.

AB#10163