fix: 10433 remove the path checking
The path checking was intended as an early guard against the case where
a mix of tenant massifs in the same directory would cause replication to
fail. The intent was to help the caller by providing them and indication
of the problematic file names

Instead, we now have tests (in veracity) which show the explicit errors
for this case are surfaced from this library and are acted on
appropriately.